chore: add fallow and apply its codebase cleanup (#187, #190)
Adds fallow (Rust-native dead-code/duplication/complexity analyzer) as a
devDependency and applies its findings:

- Demote 18 unused exports + 25 unused types to file-local; remove the
  one fully-unused function (loadExcludes was already restored — only
  truly orphaned code went)
- Extract VmHandle + dependent option/result interfaces into vm-handle.ts
- Extract runtime core (boot/attach/restore/measureFirstByte + helpers)
  out of the 1879-line index.ts into vm.ts; index.ts is now an 88-line
  façade. Breaks the index.ts <-> provision.ts cycle (#190)
- Add .fallowrc.json with entry globs for scripts/fixtures/demos and
  ignoreDependencies for string-loaded packages (node-pty, oxlint-config,
  agent-ci, typedoc-plugin-markdown)
- Suppress class-member findings on Sandboxes/Supervisor/VsockWinsize
  with rationale (fallow can't trace `new Class().method()` calls)

fallow now reports zero dead-code findings; maintainability index 93.3.
