+++ .fallowrc.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "https://raw.githubusercontent.com/fallow-rs/fallow/main/schema.json",
+  "entry": [
+    "scripts/*.ts",
+    "scripts/*.mjs",
+    "packages/runtime/src/bin/*.ts",
+    "packages/microvm/test-fixtures/assets/*.{js,ts,mjs}",
+    "examples/**/rootfs/**/*.{js,mjs,ts}"
+  ],
+  "ignoreDependencies": [
+    "@homebridge/node-pty-prebuilt-multiarch",
+    "@nkzw/oxlint-config",
+    "@redwoodjs/agent-ci",
+    "typedoc-plugin-markdown"
+  ]
+}
