+++ packages/runtime/src/errors.ts
@@ -31,6 +31,7 @@ export const ErrorCode = {
   BOOT_MODULES_NOT_FOUND: "BOOT_MODULES_NOT_FOUND",
   BOOT_CMD_WITHOUT_IMAGE: "BOOT_CMD_WITHOUT_IMAGE",
   BOOT_CMD_MISSING: "BOOT_CMD_MISSING",
+  BOOT_CWD_INVALID: "BOOT_CWD_INVALID",
   BOOT_MOUNT_INVALID: "BOOT_MOUNT_INVALID",
   BOOT_MOUNT_HOST_NOT_FOUND: "BOOT_MOUNT_HOST_NOT_FOUND",
   BOOT_PORT_FORWARD_INVALID: "BOOT_PORT_FORWARD_INVALID",
