+++ packages/runtime/src/multiplex.ts
@@ -18,7 +18,7 @@
 
 import type { Writable } from "node:stream";
 import { SandboxError } from "./errors.ts";
-import type { VmHandle } from "./index.ts";
+import type { VmHandle } from "./vm-handle.ts";
 
 export interface SandboxEntry {
   id: string;
@@ -52,6 +52,8 @@ export class Sandboxes {
     this.scrollbackBytes = opts.scrollbackBytes ?? 8 * 1024;
   }
 
+  // Public API — fallow can't trace `new Sandboxes().add(...)` calls.
+  // fallow-ignore-next-line unused-class-member
   add(id: string, vm: VmHandle): void {
     if (this.items.has(id)) {
       throw new SandboxError("SANDBOX_ID_DUPLICATE", `sandbox id already registered: ${id}`);
@@ -203,6 +205,8 @@ export class Supervisor {
   }
 
   /** Run until stopped. Resolves when input ends or stop() is called. */
+  // Public API — fallow can't trace `new Supervisor().run()` calls.
+  // fallow-ignore-next-line unused-class-member
   run(): Promise<void> {
     this.print(this.bannerText());
     return new Promise<void>((done) => {
